Adding index page tests
